+++ src/TextureManager.cpp
@@ -48,11 +48,11 @@ void TextureManager::draw(std::string id, int x, int y, int width, int height, S
     SDL_RenderCopyEx(pRenderer, m_textureMap[id], &srcRect, &destRect, 0, 0, flip);
 }
 
-void TextureManager::drawFrame(std::string id, int x, int y, int width, int height, int currentRow, int currentFrame, SDL_Renderer* pRenderer, SDL_RendererFlip flip){
+void TextureManager::drawFrame(std::string id, int x, int y, int width, int height, uint32_t currentRow, uint32_t currentFrame, SDL_Renderer* pRenderer, SDL_RendererFlip flip){
     SDL_Rect srcRect;
     SDL_Rect destRect;
-    srcRect.x = width * currentFrame;
-    srcRect.y = height * (currentRow - 1);
+    srcRect.x = width * static_cast<int>(currentFrame);
+    srcRect.y = height * static_cast<int>(currentRow - 1);
     srcRect.w = destRect.w = width;
     srcRect.h = destRect.h = height;
     destRect.x = x;
